requirements.txt
torch
torchvision
torchaudio
pandas
transformers
matplotlib
tqdm
numpy
Pillow
decord
pycocotools
pymilvus
modelscope
fastapi
uvicorn
python-multipart
aiofiles